Fix implicit 64 to 32-bit truncations in sys/ufs.
Lessons learnt: quotas will stop working in 2038, ffs1 is hopelessly
outdated and should eventually be removed, and directories can only
exist up to 2GB in size.

+++ sys/ufs/ufs/dirhash.h
@@ -118,8 +118,8 @@ void	ufsdirhash_uninit(void);
 int	ufsdirhash_build(struct inode *);
 doff_t	ufsdirhash_findfree(struct inode *, int, int *);
 doff_t	ufsdirhash_enduseful(struct inode *);
-int	ufsdirhash_lookup(struct inode *, char *, int, doff_t *, struct buf **,
-	    doff_t *);
+int	ufsdirhash_lookup(struct inode *, char *, size_t, doff_t *,
+	    struct buf **, doff_t *);
 void	ufsdirhash_newblk(struct inode *, doff_t);
 void	ufsdirhash_add(struct inode *, struct direct *, doff_t);
 void	ufsdirhash_remove(struct inode *, struct direct *, doff_t);
